Move Pallas tests to their own file
